requirements.txt
pip
setuptools
wheel
django<2.0
psycopg2-binary
django-admin-rangefilter
djangorestframework>=3.9.0,<3.10.0
pytz
coreapi
pyyaml
elasticsearch>=6.0.0,<7.0.0
neo4j
django-jet
django-debug-toolbar
django-cors-headers<3.0
gunicorn
django-rq<2.0
rq-scheduler
django-ckeditor
jsonpickle
django-otp<1.0.0
qrcode
django-recaptcha2
pycryptodome
python-slugify
django-render-block<0.7
django-templated-email
html2text
